refactor(tests): C9-08 extract run_cmd and run_python tests from test_tools_execution.py
Move run_cmd observability, timing, and headless gate tests to
test_tools_run_cmd_unit.py. Move run_python observability and headless
gate tests to test_tools_run_python.py. Split mixed-concern classes
(TestGatedToolObservability, TestHeadlessGateEnforcement) by handler.
